completed optional conflig file path flag -c implementation
