+++ app/Cargo.toml
@@ -20,6 +20,9 @@ tracing = "0.1"
 tracing-journald = "0.3.0"
 tracing-subscriber = "0.3"
 update-informer = "1.1"
+dirs-next = "2.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_yml = "0"
 
 [dev-dependencies]
 assert_cmd = "2.0"
